Expose render copy positions in the GUI and move settings to its own file
