+++ src/components/SkyBrowser/SkybrowserTabs.jsx
@@ -12,6 +12,7 @@ import Checkbox from '../common/Input/Checkbox/Checkbox';
 import TooltipSkybrowser from './TooltipSkybrowser';
 import SkybrowserFocusEntry from './SkybrowserFocusEntry';
 import styles from './SkybrowserTabs.scss';
+import SkybrowserSettings from './SkybrowserSettings.jsx'
 
 class SkybrowserTabs extends Component {
   constructor(props) {
@@ -19,18 +20,12 @@ class SkybrowserTabs extends Component {
     this.state = {
       isShowingInfoButtons: [false, false, false, false, false],
       showSettings: false,
-      showSettingsRaeInfo: false,
     };
-
     this.createTabs = this.createTabs.bind(this);
     this.createImageList = this.createImageList.bind(this);
     this.setRef = this.setRef.bind(this);
     this.showTooltip = this.showTooltip.bind(this);
     this.hideTooltip = this.hideTooltip.bind(this);
-    this.createSettings = this.createSettings.bind(this);
-    this.valueToColor = this.valueToColor.bind(this);
-    this.toggleFaceCamera = this.toggleFaceCamera.bind(this);
-    this.toggleRadiusAzimuthElevation = this.toggleRadiusAzimuthElevation.bind(this);
     this.setOpacityOfImage = this.setOpacityOfImage.bind(this);
     this.removeImageSelection = this.removeImageSelection.bind(this);
     this.createButtons = this.createButtons.bind(this);
@@ -47,14 +42,6 @@ class SkybrowserTabs extends Component {
       return { top: '0px', left: '0px' };
     }
     const { top, right } = this.infoButton.getBoundingClientRect();
-    return { top: `${top}`, left: `${right}` };
-  }
-
-  get positionRae() {
-    if (!this.raeInfo) {
-      return { top: '0px', left: '0px' };
-    }
-    const { top, right } = this.raeInfo.getBoundingClientRect();
     return { top: `${top}`, left: `${right}` };
   }
 
@@ -76,22 +63,6 @@ class SkybrowserTabs extends Component {
     isShowingInfoButtonsNew[i] = false;
     this.setState({ isShowingInfoButtonsNew });
     this.setState({ isShowingInfoButtons: isShowingInfoButtonsNew });
-  }
-
-  toggleFaceCamera() {
-    const {
-      api, isFacingCamera, selectedBrowser, selectedTarget,
-    } = this.props;
-    const uriBrowser = `ScreenSpace.${selectedBrowser}.FaceCamera`;
-    api.setPropertyValueSingle(uriBrowser, !isFacingCamera);
-  }
-
-  toggleRadiusAzimuthElevation() {
-    const {
-      api, isUsingRae, selectedBrowser, selectedTarget,
-    } = this.props;
-    const uriBrowser = `ScreenSpace.${selectedBrowser}.UseRadiusAzimuthElevation`;
-    api.setPropertyValueSingle(uriBrowser, !isUsingRae);
   }
 
   toggleShowSettings() {
@@ -306,185 +277,23 @@ class SkybrowserTabs extends Component {
     return images;
   }
 
-  valueToColor(color) {
-    return {
-      r: color[0],
-      g: color[1],
-      b: color[2],
-      a: 1.0,
-    };
-  }
-
-  createSettings(target) {
-    if (!target) {
-      return '';
-    }
-    const { isUsingRae, isFacingCamera, selectedBrowser } = this.props;
-    // Take half to display in ranges [0,1] instead of [0,2]
-    const size = target.size;
-    const colors = target.color;
-    const colorData = ['Border Color: Red', 'Green', 'Blue'];
-    const sizeData = ['Browser Width', 'Browser Height'];
-    const precisionHigh = 10;
-    const precisionLow = 2;
-    const { api } = this.props;
-    const { skybrowserApi } = this.props;
-
-    // TODO: Fix color picker
-    const colorPicker = (
-      <ColorPickerPopup
-        disableAlpha
-        color={this.valueToColor(target.color)}
-        onChange={(values) => {}}
-        placement="right"
-        disabled={false}
-      />
-    );
-
-    return (
-      <div>
-        <NumericInput
-          className="Vertical Field Of View"
-          label="Vertical Field Of View"
-          max={70}
-          min={0}
-          disabled={!skybrowserApi.setVerticalFov}
-          onValueChanged={(fov) => {
-            skybrowserApi.setVerticalFov(selectedBrowser, fov);
-          }}
-          step={1}
-          value={parseFloat(target.FOV.toFixed(precisionHigh))}
-          placeholder="value 0"
-        />
-        <Row>
-          {size.map((value, index) => (
-            <NumericInput
-              className={sizeData[index]}
-              label={sizeData[index]}
-              max={1}
-              min={0.1}
-              onValueChanged={(newValue) => {
-                size[index] = newValue * 2;
-                skybrowserApi.setScreenSpaceSize(selectedBrowser, size[0], size[1]);
-              }}
-              step={0.1}
-              value={parseFloat((value * 0.5).toFixed(precisionLow))}
-              placeholder={`value ${index}`}
-            />
-          ))}
-        </Row>
-        <Row className={styles.buttonContainer}>
-          <Button
-            onClick={() => {
-              skybrowserApi.addRenderCopy(selectedBrowser);
-            }}
-            className={styles.renderCopyButton}
-            transparent
-          >
-            Add Copy
-          </Button>
-          <Button
-            onClick={() => {
-              skybrowserApi.removeRenderCopy(selectedBrowser);
-            }}
-            className={styles.renderCopyButton}
-            transparent
-          >
-            Remove Copy
-          </Button>
-        </Row>
-        <Row>
-          <NumericInput
-            className="Right Ascension"
-            label="Right Ascension"
-            max={360}
-            min={0}
-            disabled={!skybrowserApi.setVerticalFov}
-            onValueChanged={value => skybrowserApi.setEquatorialAim(selectedBrowser, value, target.dec)
-            }
-            step={0.1}
-            value={parseFloat(target.ra.toFixed(precisionHigh))}
-            placeholder="value 1"
-          />
-          <NumericInput
-            className="Declination"
-            label="Declination"
-            max={90}
-            min={-90}
-            disabled={!skybrowserApi.setVerticalFov}
-            onValueChanged={value => skybrowserApi.setEquatorialAim(selectedBrowser, target.ra, value)
-            }
-            step={0.1}
-            value={parseFloat(target.dec.toFixed(precisionHigh))}
-            placeholder="value 2"
-          />
-        </Row>
-        <div
-          ref={this.setRef('raeInfo')}
-          onMouseEnter={() => this.setState({ showSettingsRaeInfo : true})}
-          onMouseLeave={() => this.setState({ showSettingsRaeInfo : false})}
-        >
-        <Checkbox
-          label="Use Radius Azimuth Elevation"
-          checked={isUsingRae}
-          left={false}
-          disabled={false}
-          setChecked={this.toggleRadiusAzimuthElevation}
-          wide
-          style={{ width: '100%'}}
-        />
-        {
-          this.state.showSettingsRaeInfo && (
-            <TooltipSkybrowser placement="bottom-right" style={this.positionRae}>
-              {"Using RAE coordinates is going to disable interaction with the sky browser."}
-            </TooltipSkybrowser>
-          )
-        }
-        </div>
-        <Checkbox
-          label="Face Camera"
-          checked={isFacingCamera}
-          left={false}
-          disabled={false}
-          setChecked={this.toggleFaceCamera}
-          wide
-        />
-        <Row className={`${styles.vectorProperty} ${this.disabled ? styles.disabled : ''}`}>
-          {colors.map((color, index) => (
-            <NumericInput
-              className={colorData[index]}
-              label={colorData[index]}
-              max={255}
-              min={0}
-              onValueChanged={(value) => {
-                const newColor = colors;
-                newColor[index] = value;
-                skybrowserApi.setBorderColor(
-                  selectedBrowser,
-                  newColor[0],
-                  newColor[1],
-                  newColor[2],
-                );
-              }}
-              step={1}
-              value={color}
-              placeholder={`value ${index}`}
-            />
-          ))}
-        </Row>
-      </div>
-    );
-  }
-
   render() {
     const {
-      data, maxHeight, minHeight, targets, selectedBrowser, height, skybrowserApi
+      data, maxHeight, minHeight, targets, selectedBrowser, height, api, skybrowserApi, isUsingRae, isFacingCamera
     } = this.props;
     const { showSettings } = this.state;
 
     let tabDisplay = "";
     if (showSettings) {
-      tabDisplay = this.createSettings(targets[selectedBrowser]);
+      tabDisplay = (
+        <SkybrowserSettings
+          target={targets[selectedBrowser]}
+          isUsingRae={isUsingRae}
+          isFacingCamera={isFacingCamera}
+          selectedBrowser={selectedBrowser}
+          api={api}
+          skybrowserApi={skybrowserApi}
+        />);
     } else if (data.length === 0) {
       tabDisplay = <CenteredLabel>There are no selected images in this sky browser.</CenteredLabel>;
     } else {
